Moved test to s2n_client_cert_verify_test.c based on PR feedback. Resolved a potential race condition by moving apply into the main thread. Fixed remaining memory leaks.
